+++ docs/dev_docs/changelog.md
@@ -10,6 +10,7 @@
 
 * What was formerly `taylor_mode()`, is now `taylor_mode_scan()` and stands in contrast to the new `taylor_mode_unroll()`.
 * What was formerly `forward_mode()`, is now `forward_mode_recursive()`.
+* The entire `taylor` subpackage moved to top-level. Instead of `from probdiffeq.solvers.taylor import ...`, use `from probdiffeq.taylor import ...`. 
 
 
 ## v0.2.2